// apps/web-v4/src/app/models/certification.model.ts
export interface ICertificationMeta {
  isCompleted: boolean;
  warningMessage: string;
  eligibility: {
    eligible: boolean;
    reason: string;
  };
  lastTakenDate: number;
  slotFreezeDate: number;
  booking?: ICertificationBooking;
  verification_request?: ICertificationVerificationRequest;
  budget_request?: ICertificationBudgetRequest;
}

export interface ICertificationBooking {
  booking_type: string;
  dc?: string;
  testcenter?: string;
  country?: string;
  location?: string;
  date: number;
  slot: string;
  slotno: number;
  status: string;
  key?: string;
  icfdId?: number;
}

export interface ICertificationVerificationRequest {
  exam_date: number;
  result_type: string;
  result: string;
  verifierEmail: string;
  document: {
    document_url: string;
    document_name: string;
  }[];
  status: string;
  reject_reason: string;
}

export interface ICertificationBudgetRequest {
  manager_id?: string;
  amount: number;
  currency: string;
  status: string;
  reject_reason: string;
}
// ...
export type TCertificationView = 'default' | 'acc' | 'atDesk' | 'budget' | 'upload';